% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conexao.R
\name{vigiar_desconectar}
\alias{vigiar_desconectar}
\title{Disconnect and clear VIGIAR session}
\usage{
vigiar_desconectar()
}
\value{
Invisibly, \code{NULL}.
}
\description{
Disconnect and clear VIGIAR session
}
