% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{vigiar_rj_municipios}
\alias{vigiar_rj_municipios}
\title{List all 92 Rio de Janeiro municipalities}
\usage{
vigiar_rj_municipios()
}
\value{
A data.frame with columns: codigo_ibge, municipio,
macrorregiao_saude, regiao_saude.
}
\description{
List all 92 Rio de Janeiro municipalities
}
