% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{vigiar-package}
\alias{vigiar}
\alias{vigiar-package}
\title{vigiar: VIGIAR Environmental Health Data for Rio de Janeiro}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

O pacote \strong{vigiar} baixa, processa, valida e diagnostica dados de
vigilancia em saude ambiental do dashboard VIGIAR do Ministerio da
Saude, com foco no estado do Rio de Janeiro. Consulte
\code{vigiar_lifecycle()} para ver o status individual de cada funcao.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/santosry/vigiar}
  \item \url{https://santosry.github.io/vigiar/}
  \item Report bugs at \url{https://github.com/santosry/vigiar/issues}
}

}
\author{
\strong{Maintainer}: Ryan Santos \email{ryan.paulo@gsuite.iff.edu.br} (\href{https://orcid.org/0009-0005-6770-2001}{ORCID})

Authors:
\itemize{
  \item Ryan Santos \email{ryan.paulo@gsuite.iff.edu.br} (\href{https://orcid.org/0009-0005-6770-2001}{ORCID})
}

}
\keyword{internal}
